src: Make index of header int16_t

+++ src/http2.h
@@ -214,7 +214,7 @@ enum {
   HD_MAXIDX,
 };
 
-using HeaderIndex = std::array<int, HD_MAXIDX>;
+using HeaderIndex = std::array<int16_t, HD_MAXIDX>;
 
 // Looks up header token for header name |name| of length |namelen|.
 // Only headers we are interested in are tokenized.  If header name
